Define project manager duties when removing messages
I think we should ensure that project managers cannot just randomly delete things. As a non native speaker i have a hard time phrasing this correctly, but basically the deletion needs to be traceable by other project managers. Without having a clear example, I have seen in the past that in other projects it could be come an issue if a project manager gets into a heated discussion wiht someone else and deletes comments without the ability for others to review the process.